Bugfix in conf.sh
When training from scratch, the tokenizer did not get passed downstream.

This issue was not noticed before due to our focus on continued training, but now I was unable to start training without this fix.

Error msg:
```python
ValueError: You are instantiating a new tokenizer from scratch. This is not supported by this script.You can do it from another script, save it, and load it here, using --tokenizer_name.
```

Tagging @RikVN, @nljubesi to inform you of this bug, no further action required.